# Reduce contracts and deployments loading time