--- a/packages/playground/src/utils/batch_process.ts
+++ b/packages/playground/src/utils/batch_process.ts
@@ -1,0 +1,13 @@
+export async function batchProcess<T, R>(
+  items: T[],
+  batchSize: number,
+  processFn: (batch: T[]) => Promise<R[]>,
+): Promise<R[]> {
+  const results: R[] = [];
+  for (let i = 0; i < items.length; i += batchSize) {
+    const batch = items.slice(i, i + batchSize);
+    const batchResults = await processFn(batch);
+    results.push(...batchResults);
+  }
+  return results;
+}