@0oM4R 2025-06-23
there are other places please make sure to check them as well and replace the promise.all with promise.allsettled