@0oM4R 2025-07-15
I tried to make the deployment listing fail 

The error is not accurate, imo
<img width="1838" height="782" alt="Pasted Graphic 2" src="https://github.com/user-attachments/assets/e9ef1261-b70e-4f85-bddb-4be031fee918" />
<img width="1815" height="885" alt="Pasted Graphic 1" src="https://github.com/user-attachments/assets/13e0f1df-8ea1-4f55-a07d-1cc6c2fb3976" />