@ramezsaeed 2025-07-17
Tested:
 - Loading time enhanced with 50 VM contracts