@samaradel 2025-07-20
@0oM4R I see the pagination error exists in development, for the contracts I tested with more than 50 contracts, and it takes 13 seconds